Added combat list page
Added combat list page
