+++ TimeForBattle/TimeForBattle/AppShell.xaml.cs
@@ -7,6 +7,7 @@ public partial class AppShell : Shell
     public AppShell()
     {
         InitializeComponent();
+        Routing.RegisterRoute(nameof(CombatListPage), typeof(CombatListPage));
         Routing.RegisterRoute(nameof(CreatureListPage), typeof(CreatureListPage));
         Routing.RegisterRoute(nameof(CreatureDetailsPage), typeof(CreatureDetailsPage));
         Routing.RegisterRoute(nameof(InitiativePage), typeof(InitiativePage));
