+++ TimeForBattle/TimeForBattle/View/CombatListPage.xaml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:model="clr-namespace:TimeForBattle.Model"
+             xmlns:viewmodel="clr-namespace:TimeForBattle.ViewModel"
+             xmlns:converters="clr-namespace:TimeForBattle.Converters"
+             x:DataType="viewmodel:CombatListViewModel"
+             x:Class="TimeForBattle.View.CombatListPage"
+             Title="{Binding Title}">
+
+    <ScrollView>
+        <FlexLayout BackgroundColor="Transparent"
+                    BindableLayout.ItemsSource="{Binding Combats}"
+                    Direction="Column"
+                    Wrap="NoWrap"
+                    JustifyContent="Start"
+                    AlignContent="Start">
+            <BindableLayout.ItemTemplate>
+                <DataTemplate x:DataType="model:Combat">
+                    <Border Stroke="DarkRed"
+                            StrokeThickness="2">
+                        <Grid ColumnDefinitions="*,*,*"
+                              RowDefinitions="*,*">
+                            <Label Grid.Column="0"
+                                   Grid.ColumnSpan="2"
+                                   Grid.Row="0"
+                                   Grid.RowSpan="2"
+                                   Text="{Binding Name}"
+                                   FontSize="20"
+                                   FontAttributes="Bold"
+                                   HorizontalOptions="Start"
+                                   VerticalOptions="Start"
+                                   Margin="5,5,0,0" />
+                            <Button Grid.Column="2"
+                                    Grid.Row="0"
+                                    Text="Load"
+                                    CommandParameter="{Binding .}"
+                                    Command="{Binding x:DataType='viewmodel:CombatListViewModel', Source={RelativeSource AncestorType={x:Type viewmodel:CombatListViewModel}}, Path=LoadCombatCommand}" />
+                            <Button Grid.Column="2"
+                                    Grid.Row="1"
+                                    Text="Delete"
+                                    CommandParameter="{Binding .}"
+                                    Command="{Binding x:DataType='viewmodel:CombatListViewModel', Source={RelativeSource AncestorType={x:Type viewmodel:CombatListViewModel}}, Path=DeleteCombatCommand}" />
+                        </Grid>
+                    </Border>
+                </DataTemplate>
+            </BindableLayout.ItemTemplate>
+        </FlexLayout>
+    </ScrollView>
+</ContentPage>
