+++ TimeForBattle/TimeForBattle/ViewModel/MainMenuViewModel.cs
@@ -7,26 +7,13 @@ public partial class MainMenuViewModel : BaseViewModel
 {
     public CreatureService<Combat> CombatService;
     public ObservableCollection<Combat> Combats { get; }
+    [ObservableProperty] bool isRefreshing;
 
     public MainMenuViewModel(CreatureService<Combat> combatService)
     {
         Title = "Main Menu";
         this.CombatService = combatService;
         Combats = [];
-    }
-
-    [ObservableProperty] bool isRefreshing;
-
-    [RelayCommand]
-    public async Task RefreshCombats()
-    {
-        List<Combat> combatData = await CombatService.GetAllAsync();
-        Combats.Clear();
-
-        foreach (Combat combat in combatData)
-        {
-            Combats.Add(combat);
-        }
     }
 
     [RelayCommand]
@@ -38,6 +25,7 @@ public partial class MainMenuViewModel : BaseViewModel
         await CombatService.SaveAsync(combat);
 
         combat.Name = "Combat #" + combat.Id.ToString();
+
         await CombatService.SaveAsync(combat);
 
         await Shell.Current.GoToAsync($"{nameof(InitiativePage)}", true,
@@ -48,28 +36,8 @@ public partial class MainMenuViewModel : BaseViewModel
     }
 
     [RelayCommand]
-    public async Task DeleteCombatAsync(Combat combat)
+    public async Task GoToCombatListPageAsync()
     {
-        if (combat is null)
-            return;
-
-        await CombatService.DeleteAsync(combat);
-
-        //also delete creatures in initiative for that combat?
-
-        await RefreshCombats();
-    }
-
-    [RelayCommand]
-    public async Task LoadCombatAsync(Combat combat)
-    {
-        if (combat is null)
-            return;
-
-        await Shell.Current.GoToAsync($"{nameof(InitiativePage)}", true,
-            new Dictionary<string, object>
-            {
-                {"Combat", combat}
-            });
+        await Shell.Current.GoToAsync($"{nameof(CombatListPage)}", true);
     }
 }
